requirements.txt
matplotlib==3.7.3
numpy==1.26.1
pandas==2.0.3
PyYAML==6.0.1
scikit_learn==1.3.0
torch==2.1.0
